requirements.txt
Django==4.2.7
django-allauth==0.58.2
django-background-tasks==1.2.5
django-debug-toolbar==4.2.0
django-htmx==1.17.2
django-template-partials==23.3.post1
django-ordered-model==3.7.4
django-simple-history==3.4.0
django-ninja==1.1.0
django-taggit==5.0.1
django-filter==23.5
django-extensions==3.1.3
django-invitations
django-split-settings==1.2.0
django-widget-tweaks==1.5.0
factryengine
gunicorn==22.0.0
numpy==1.26.4
pillow==10.2.0
pyparsing==3.1.1
python-dotenv==1.0.0
psycopg2==2.9.9
pydot==2.0.0
graphviz==0.20.1
django-role-permissions==3.2.0
django-storages[azure]==1.13.2
whitenoise
